assignments/week4/LangChain_project/simple_genapp/app.py
from langchain import OpenAI, LLMChain, PromptTemplate

import os
from dotenv import load_dotenv

load_dotenv()
llm = OpenAI(api_key=os.getenv("ΟΡΕΝΑΙ_API_KEY"),temperature=0.7)
prompt = PromptTemplate(
    input_variables=["user_input"],
    template="""
    You are a helpful AI assistant.
    User says: {user_input}
    Your response:
    """
    )
chain = LLMChain(llm=llm, prompt=prompt)

if __name__ == "__main__":
    user_input = input("Ask me anything: ")
    response = chain.run({"user_input": user_input})
    print("AI says:", response)